requirements.txt
altgraph==0.17.4
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
comm==0.1.4
darkdetect==0.7.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
flake8==6.0.0
fqdn==1.5.1
h11==0.14.0
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
jedi==0.19.0
jeepney==0.8.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.1
MouseInfo==0.1.3
mss==9.0.1
mypy==1.4.1
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.4
notebook_shim==0.2.3
openpyxl==3.1.2
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyGetWindow==0.0.9
Pygments==2.16.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
pyqtdarktheme==2.1.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-xlib==0.15
pytweening==1.0.7
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
selenium==4.12.0
Send2Trash==1.8.2
shiboken6==6.5.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
stubs==1.0.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.10.1
trio==0.22.2
trio-websocket==0.10.4
types-requests==2.31.0.5
types-urllib3==1.26.25.14
uri-template==1.3.0
urllib3==2.0.5
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
wsproto==1.2.0
